// src/app/models/company.model.ts
// This is the model for the company info
export interface CompanyI {
    uid: string;
    name: string;
    description: string;
    logo: string;
    email: string;
    address: string;
    country: string;
    city: string;
    state: string;
    zip: string;
    phone: string;
    fax: string;
    website: string;
    facebook: string;
    twitter: string;
    instagram: string;
}